refactor: separate component responsible for managing notes - separate component in a different file
