+++ src/windows/note/styles.css
@@ -1,0 +1,57 @@
+@import "tailwindcss";
+@import "tw-animate-css";
+
+.card {
+  background-color: var(--color-card-bg);
+  @apply overflow-hidden w-dvw h-dvh;
+}
+
+.card>header {
+  background-color: var(--color-black-1);
+  @apply h-[32px] w-full flex justify-between items-center p-2;
+  -webkit-app-region: drag;
+}
+
+button {
+  -webkit-app-region: none;
+}
+
+.card>header>.close-and-elpse-button {
+  @apply flex gap-4;
+}
+
+.menu {
+  @apply bg-white flex flex-col gap-4 items-center absolute right-[3px] top-8 px-2 py-4 rounded-b-lg;
+}
+
+.menu>li>button {
+  @apply flex gap-1 justify-center;
+}
+
+
+.menu>li>.credits {
+  @apply flex gap-1 flex-col items-center text-sm;
+}
+
+.menu>li>.credits>p>a {
+  @apply text-blue-700;
+}
+
+.menu--disabled {
+  transform: translate(0, -100vh);
+  @apply overflow-hidden ease-in -z-10;
+}
+
+.menu--disabled {
+  transform: translate(0);
+  @apply duration-500 ease-in;
+}
+
+.card>textarea {
+  @apply resize-none w-full border-0 outline-none p-3 overflow-auto;
+  height: calc(100% - 68px);
+}
+
+.card>footer {
+  @apply bg-white flex justify-between p-1 outline-1 outline-[#141414];
+}
